Add tests for int, bool, and double
