+++ tests/run-macros/annot-arg-value-in-java/Annot.java
@@ -1,0 +1,5 @@
+public @interface Annot {
+  int value() default 0;
+  boolean m() default false;
+  double n() default 0;
+}
