+++ tests/run-macros/annot-arg-value-in-java/S.scala
@@ -1,0 +1,9 @@
+// Display annotation arguments in Java
+
+// inline def showAnnots(inline c: String): Unit =
+//   println(c + ":")
+//   annots(c).foreach(println)
+
+@main def Test =
+  showAnnots("J")
+  showAnnots("JOtherTypes")
