--- tests/run/annot-arg-value-in-java/AnnoMacro.scala
@@ -1,7 +1,0 @@
-import scala.quoted.*
-
-inline def annots(inline c: String): List[String] = ${ annotsImpl('c) }
-
-def annotsImpl(c: Expr[String])(using Quotes): Expr[List[String]] =
-  import quotes.reflect.*
-  Expr(Symbol.requiredClass(c.valueOrError).declaredMethods.map(_.annotations.toString))
