--- tests/run/annot-arg-value-in-java/J.java
@@ -1,8 +1,0 @@
-public class J {
-  @SuppressWarnings(value = "a")
-  public void f1() {}
-  @SuppressWarnings("b")
-  public void f2() {}
-  @SuppressWarnings({"c", "d"})
-  public void f3() {}
-}
