--- tests/run/annot-arg-value-in-java/S.scala
@@ -1,4 +1,0 @@
-@main def Test =
-  inline val c = "J"
-  println(c + ":")
-  annots(c).foreach(println)
